Add fixed top bar and separate menu toggle JS
Moved the Javascript for the menu toggle into it's own separate file.
Also added fixed-top-bar.css for those wishing to have the menu toggles
and branding fixed to the top of the page. The later is enabled by
default, however if you would like it disabling simply remove the link
to the stylesheet from the index.html file.